maintainability: ratchet reconciliation and accumulated local edits
- maintainability_metrics.py: functions_over_100 budget 21->19 (genuine
  reduction from helper extraction); regenerate ty baseline.
- renderers/diff.py crossed 500 LOC via helper extraction (cohesive
  renderer, recorded as accepted_hotspot).
- captures other in-flight local working-tree edits.

+++ src/quant_execution_engine/execution_policy.py
@@ -150,7 +150,7 @@ def discretize_quantity(quantity: float, *, lot_size: int) -> int:
         raise ValueError("lot_size must be > 0")
     if not math.isfinite(float(quantity)):
         raise ValueError("quantity must be finite")
-    return int(round(float(quantity) / lot_size)) * lot_size
+    return round(float(quantity) / lot_size) * lot_size
 
 
 def participation_capped_quantity(
@@ -195,7 +195,7 @@ def build_dynamic_execution_decision(
         market_price=market,
         omega=omega,
     )
-    raw_target = int(round(normalized_target * config.max_position))
+    raw_target = round(normalized_target * config.max_position)
     target = discretize_quantity(raw_target, lot_size=config.lot_size)
     target = max(-config.max_position, min(config.max_position, target))
     order = target - int(current_quantity)
